+++ CHANGELOG.md
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated colors in the app to be accessible to colorblind people using the [Bang Wong](https://www.nature.com/articles/nmeth.1618) palette.
 - Update the file that stores taxa_names to use the common names in the drop-down menu.
 - Update the title and description in index.html.
+    - A drawing tool allows users to select a rectangle or polygon on the map.
+    - Users can toggle to add or remove hexagons (presence or absence).
+    - The selected area is automatically filled with or cleared of hexagons based on the chosen type.
 
 ### Fixed
 - The `Taxa Name` info box now correctly clears when an invalid taxa ID is entered.